requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
feedparser==6.0.11
httpx==0.27.2
lxml==5.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.35
typer==0.12.5

